requirements.txt
certifi==2021.10.8
cycler==0.11.0
h5py==3.2.1
joblib==1.1.0
kiwisolver==1.3.2
matplotlib==3.4.3
numpy==1.21.5
pandas==1.3.5
Pillow==8.0.1
pyparsing==3.0.6
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
scikit-learn==1.0.1
scipy==1.7.3
seaborn==0.11.2
six==1.16.0
threadpoolctl==3.0.0
torch==1.4.0
torchvision==0.5.0
tqdm==4.43.0
